add deployed url section
