+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/kayleriegerpatton/readme-generator#readme",
   "dependencies": {
+    "colors": "^1.4.0",
     "fs": "^0.0.1-security",
     "inquirer": "^8.2.0"
   }
